+++ src/toisto/command/__init__.py
@@ -1,4 +1,5 @@
 """Commands."""
 
 from .practice import practice
+from .show_topics import show_topics
 from .show_progress import show_progress